Add Category model
Adds a category model, uniqueness constraint, relevant tests.

This also adds an entity to setup how we want to return the category
data in the app. In the UI, we'll want to render all the category and
subcategory data together and know if they can be deleted or not. The
category_data entity takes care of all of that for us.

The data that relies on subcategory and transaction is mocked out in
the entity for now. As soon as we add those models and associations,
the entity logic will be filled in.