Add Subcategory model
This commit adds the subcategory model which belongs to a category.

We also update the category_data entity to return the correct
subcategory data.

Routes are added for mutating only: update, create, destroy. We don't
need any routes for showing the data because the application will
display them along with the categories, and therefore, we use category
index.