Add Transaction Model
Adds the transaction model and all associations. The transaction
can belong to a statement, or just an account if added manually.
Most of the transactions will be added via an automated background
job that parses monthly account statements.

The Category and Subcategory are updated to not allow deletion if
they have associated transactions.