Refine behavior for transaction categories
Refactors the create and update methods so that the transaction's
category and subcategory are specified by name, not by ids. This makes
the user experience easier by not having to worry about category or
subcategory IDs.

Adds logic to handle default values for category and subcategory
for new transactions. If a category or subcategory isn't specified,
then it defaults to uncategorized.

There is also logic to return a helpful error if an invalid category
is passed in for the create or update methods.

Tests were added for these edge cases.