Further refine transaction category associations
This updates the transaction controller to no longer accept category
for creating or updating a transaction. Only the subcategory is necessary
and the category will be set based on the subcategory.

We also update the subcategory param to subcategory_name to reflect the value
expected more accurately.

This also enables Action Controller Params Wrapper subcategory_name.
This means that if a user passes in this attribute and it is not
nested under the "transaction" object, Rails will do it for us
automatically.

Because of this, we are able to remove the "transaction" object from the
controller tests. We also add tests to ensure the category and subcategory
were set correctly.