Add Transaction serializer
Adds a new serializer for rendering a single transaction. The
Category and Subcategory association will be rendered with the name
rather than the id.

This is different from the Transaction Data Entity in that the entity
is for rendering all of the transactions and associated data, and the
serializer is for rendering a single transaction.

Affected tests were updated to handle the new attributes.