Add category data entity
Adds a new entity for displaying all of the categories and
subcategories. The entity is the only way to retrieve this data as
the show endpoints for both models don't exist. This is okay because
the number of categories and subcategories won't be high, and all of
them can be returned. In fact, im thinking about setting a hard limit
of 500 categories and 1000 subcategories. That is not in this commit,
but might come later.