Initial commit of frontend
Adds the frontend app, which is a next.js react app. This initial
commit includes the layout and the transactions table.

The ability to update subcategories on a transaction is working as
well.

Unit tests were not added yet as I haven't learned how to do them. I
will add them in a future commit once I'm ready to learn how to write
frontend tests.