Refactor transaction serializer
Refactors the transaction serializer so that the response of the
create and update actions return the same structure as the index.

This simplifies the frontend logic so that when we update a transaction,
we can easily update the transactions state with the same attributes.

Spec tests were updated to use the new return values.