Refactor Transaction Data Entity
Update the transaciton data entity to make the logic more clear.

The accounts and users filters do an include instead of an exclude.
The UI will be updated to accommodate this new logic which is okay
since the ability to filter on these in the UI isn't avilable yet.

New params were added to allow sorting by transaction_date or amount,
ascending or descending with the sort_by and sort_direction params.

The query param was renamed to search_string to make it more clear
and reduce confusion with the ActiveRecord query "query" variable.

A lot of unit tests were backfilled for the transaction data entity.