Implement ending_before option
Implements the ending_before option within the transaction_data_entity.
This option allows users and the UI to navigate to the previous page of
data.

The next_page and prev_page cursors were also implemented to make it
easy to jump to the next and previous page.

Extensive spec test coverage is added to ensure every edge case is
covered.