Refactor Transactions page
This commit completely refactors the Transactions Page format and
logic. The first update is the format of the page. It is now split
into three main columns, with the center column being the main
content. The left column is hidden on screen sizes "md" and smaller.
When the screen becomes larger than "md", it will display the left
column and the filter bar will move into that space.

The search bar and create transaction will always remain on top of
the table.

The table columns were reduced to only show Date, Description,
Category, Amount and an icon to hint at expanding the row.

Functions were added to fetch users and accounts and the data stubs
were removed.

The filter dropdowns were refactored to not use checkboxes but rather
allowing a click on the item itself. The filters were udpated to filter
for rather than exclude the account. An "All" option was added to each
dropdown to clear the filter.