Refactor the transaction expanded row
This refactors the transaction expanded row so that it neatly
"expands" the row to show more transaction details. I quote "expands"
because what's actually happening is another row is added under
the selected transaction with a cell that spans all 5 columns.

This also updates the collapse button so that it's in the expanded
cell, and it takes up the full height.

All of the cells between the TransactionTableRow and
TransactionTableRowExpanded were udpated so that when the row is
expanded, nothing shifts around.

All items are now aligned to the top of the row.

There is a current bug where if the row is expanded, the Account
Filter dropdown will leak the date cell into the dropdown. That
will be fixed in a future commit, or perhaps we just collapse
the row when anything outside the row is clicked.