Update filter dropdowns
This updates the filter dropdowns to adds visual effects to the
action buttons to make it look like they are pressed when a user
clicks on them.

This also updates the param name to include array brackets: [].
This is how the browser sends array query params to the backend.
Prior to this change, the backend was only processing the last
value in the query array param.

And finally, we add a ref to each of the dropdowns with listeners
to detect if the user has clicked off of them. If so, the dropdowns
will close.