Extract category dropdown item
Extracts the category dropdown item into its own component. Also
improves styling so that the list item is clickable instead of using
an inner div. This caused issues when clicking because the list item
had padding, which wasn't clickable. Now the click zone of the
subcategory is easier to target.