Update transaction_date and date picker
Updates the date picker so that it's popper placement is bottom-end.
This allows it to show underneath the date field when expanded and
without cutting it off of the left side of the screen.

This also reduces the width of the date field to fit within the column.
The DatePicker component was pushing the width of the column out.

Additionally, we update the backend to return the transaction date in
"mm/dd/yyyy" format.