Rework transaction expanded row close listener
Rework the transaction expanded row close listener so that the
row collapses when clicked outside of the expanded row. This also
adds logic to check if the click is inside another row of the
TransactionsTable. If it is, we don't want two functions to collapse
the row because it behaves in an unpredictable way. Instead, if a
different row is clicked, we simply swap the expanded row for this
new row. However if the click is outside of the table, then we
collapse the row and don't expand another one.

This also adds some more details to the expanded row pane such as
original statement description, account and a split transaction
span which will be turned into a button to split the transaction.