Clean up console warnings for logo
Adds the style width: auto height: auto to the image tag of the logo.
Next.JS was complaining that the aspect ratio could be skewed so
this resolves that warning.

Another warning was around not having a priority tag since the logo
is a largest contentful paint element. The priority tag will preload
the image to speed up LCP.