Add columns to Transaction
This adds 3 columns to the Transaction model:
1. notes
2. statement_description
3. statement_transaction_date

This also modifies the existing columns to set not nullable:
1. transaction_date
2. amount

statement_description will be the original description from the
bank statement and is unchangable once set. The description is
allowed to be modified to the users liking. Same thing goes for
the statement_transaction_date column since transaction_date is
allowed to be modified.

All annotations have been updated.