Update transaction response
Adds the new fields to the transactions API response as well as
updates the typescript definition.

I also undid the date formatting on the backend. The formatting will
need to be completed on the frontend, which will be more efficient
and not break the pagination cursor, which is a combination of date+id.