Add columns to accounts and statements
Adds a column for specifying the statement directory of an account.
The plan is to use the directory path from the Documents/Financial/
directory to look for the PDF statements of an account to parse.

A new column was added to statement model for pdf_file_path, which
will record the file path when the statement is getting processed.
This will help check if a statement was processed or not while the
parser job is running.