Set Uncategorized as default subcategory
Modifies the transaction model logic to set the default category
to "Uncategorized" if no subcategory is specified. This also
updates the logic in the transaction model to set the category
based on the subcategory.

This change allowed us to simplify the logic in the Transaction
Controller: we don't need to lookup the category from the subcategory
in the controller anymore, and we don't need to set the default value
there either. All of this now happens in the model.

I would add tests for this, however, minitest keeps crashing with
a segfault when we run "rails test". As soon as that's fixed we'll
add tests that we can make sure work.