Add statement_parser class base
Adds a new statement_parser class that will be used as the abstract
class for all of the bank's parser classes. Each bank statement will
have its own dedicated parser_class which will inherit from the base.

Each parser class must implement the following methods:
- statement_end_date
- statement_start_date
- statement_balance
- transactions

In a followup commit, a Job will be added that iterates each account,
and uses the account's parser class to extract necessary data from
the statement PDF.

This commit also adds a few new Gems that will be necessary for the
parser background job such as pdf-reader, as well as testing utilities
such as timecop and mocha.