Add parser_class to account
Adds a new parser_class attribute to account. This column will store
the class that is needed to parse the statements for this account.

An instance method is added to the account model to instantiate a
new parser_class with the statement PDF for that account. This will
eventually be called by a background job that is responsible for
parsing and storing the transactions from a bank statement.