Add a Parse Statements Job
This adds a Parse Statements Job that will be responsible for
parsing unprocessed account statements, retrieving the data, and
storing it in the database. It will create new statement records as
well as transaction records.

Accounts must have the statement_directory and parser_class attributes
populated in order for its statements to be processed.

The ROOT_DOC_DIR is the root directory where all of the statements
are stored. This will be different in devel vs prod, so an ENV
variable seems fitting.