Add Schwab Checking statement parser
This commit adds the parser class for Schwab Checking statements.
This one was very tricky because some transaction descriptions can
span multiple lines. The debit, credit belance columns are based on
whitespace. The amounts don't use negative numbers to signal credits
or debits, you need to determine which column the value is in.

It is also possible that whitespace counts overlap between debits
and credits, so we can't rely on whitespace counts to determine
which column the amount is under. Therefore, we use keywords to
determine if an amount is a debit or credit.

Precautions were taken to raise an error if a transaction matches
keywords from the debit and credit list, as well as if it doesn't
match any of the keywords.

Additionally, the statement date range varies based on if the
statement is contained within a single month, or if the days span
multiple months.

I feel strongly that I wrote test cases to cover all conditions, and
now that they are passing, I believe this parser is ready.