Refactor Transaction Meta Data
Adds a new meta object under the root ojbect of the transaction data
entity. The meta object contains all of the data related to fetching
the transactions such as item counts, and pagination tokens.

This also updates the Typescript definitions to match the response.

We also take this opportunity to store the TransactionMetaData items
upon successful fetch.

All tests were updated to adopt this new shape.