Format transaction table data
This adds color formatting to the amount based on if it's negative
or positive. Negative amounts are printed in red while positive ones
are printed in green.

This also updates the transaction date to be mm-dd-yyyy instead of
keeping the backend format of yyyy-mm-dd. We want to keep the backend
in the same format that it accepts the dates in when updating or
creating a transaction record. Therefore, the UI will do the
formatting.