Fix Transaction DatePicker bug
This fixes a bug where the DatePicker would display the date as one
day earlier than the actual transaction date. The reason was due to
timezone issues. Using the Javascript date-fns library, we can create
a new date object that doesn't have time. This allows the DatePicker
to select the correct date.