Fix layout and nav bar
This updates the transaction page layout to remove the right pane
of the page layout. This was turning out to get in the way of the
clickable components on the table. I'm still unsure why the
TransactionTable was being pushed behind this empty pane, but we
dont need the pane, so removing it simlifies the UI.

Additionally, we remove the logic to move the filter bar to
the left pane when the responsive size is lg or higher. It
was causing issues and isn't necessary right now. We leave
it simple for now, and can add more complex functionality later.

This also updates the Navlinks to use the Inter font, and the Tincan
name to use Abril Fatface.