Add ability to update description
Adds the ability to update the description of a transaction. When
the transaction row is expanded, the description will be a text input
field with a save button.

I added some user guidance to make the input field border red when
the field isn't saved. The button also updates to "Saved" when the
description input field is in sync with the transaction description.

Also, an onKeyDown handler was added so that if the user presses
enter/return when the text input has focus, it will save the input.
We added a check to ensure that we only make an XHR if the description
needs to be updated. Otherwise accidentally holding down the enter key
will send a lot of XHRs to update the description.