Fix date bug in Safari
This fixes a bug that would cause the website to crash in Safari.
Safari couldn't handle:
new Date(formattedDate)

We removed this line and confirmed that parseISO was sufficient to
get the DatePicker to work.

This also takes the opportunity to do some small fixups. One fixup
is reducing the padding on the sides of the date picker input and
description text input. This allows all of the date to be seen when
the window is md or smaller.

Additionally, we add a border around the date picker input so that it
now matches the description input box.

Also removes a leftover debug line from a previous commit.