Add Chase Freedom Statement Parser
Adds the Chase Freedom Statement Parser class and all relevant
tests.

This also fixes a weird typo in the BarclaysView statement parser
class where we weren't selecting the statement_end_date's year in
the calculate_transaction_year method. The weird part about this
was that the code still worked which was just very lucky.