Redesign the transaction filter column
Adjusts the filter bar to be a column on the left side of the
transactions table. This also updates margins within the layout,
transactions page and filter dropdowns so that they are correctly
spaced.

The biggest change here is that the dropdowns are no longer absolute
positioned. Expanding them will cause the contents below the button
to drop down. Since the filter bar is no longer directly above the
table, we can allow the contents to be pushed down because it won't
change the entire page.