Add generalized filter selector component
This commit creates a new generalized filter selector component
by copying common logic from the UserSelector and the
AccountSelector.

This new component will be able to take in a title and a list
of items which contain an id, label and optional sublabel.
The Filter Selector will display these items properly when clicking
the button. The button will have the title as a label.

A onItemClick function is also passed in as a prop to handle when a
user clicks on one of the filter items.

A selectedItems array is passed in to the component to allow us to
differentiate the background of the selected items.

At this time, this new filterSelector component is not being utilized.
A followup commit will refactor the filterbar component to use it.