Refactor transaction filter to use FilterSelector
This commit renames the filterbar to TransactionFilter. It also
updates the logic of this higher order component to use the new
generalized FilterSelector component.

The onClick logic was extracted from the old, individual filter
selectors up into this higher order component.

The old AccountSelector and UserSelector componenents are no longer
utilized and will be removed in a followup commit.