Remove old filter selectors
Deletes the old UserSelector and AccountSelector components. These
are no longer used as of the last commit.

Additionally, Kate wanted the users filter to be above the account
filter on the transaction page, so this swaps them.

We also move the Search and button components out of the table
subdirectory because they aren't actually part of the table.

Also takes the opportunity to reorder some imports on the transaction
page so that the official libraries come first.