Update filter selector Item ID to string
This updates the FilterItemType.id to be a type string. We need to
make this a string in order to make the filter selector more versatile.

We are running into a use case where the item ids need to be non-int
values. This change allows us to utilize the filterSelector component
for this use case.