Fill in Account page with table
This fills in the Account page with a table that displays each bank,
and each bank will group the accounts together. The table is sorted
by bank name.

If there are multiple users registered in the system, the table will
be sorted by bank name and then user id. So if two users have the same
bank, they will be displayed next to each other, but sorted by the
user id for consistency.

This also removes the parserClass attribute from the Account type
definition. We don't need to return this and in a future commit, the
Account serializer will be updated to remove this attribute.