Add AccountSubRow component and Account modal
Extracts the account specific visuals from the BankRow into an
AccountSubRow component. This AccountSubRow component will be
repsonsible for rendering clickable tiles for each account, and
making a modal appear/disappear when necessary.

The account modal is added in this commit as well. It's just a
skeleton showing all of the attributes right now. In a future commit
it will be updated to handle editing some fields, deleting the account,
and marking the account as inactive.