Add the supported accounts endpoint
This adds a new endpoint to return the application's supported
accounts. This list is driven from the statement parser classes
because in order to support an account, there must be a parser
class for it.

This also adds the necessary constants to every existing parser
class that exists today, such as bank_name, account_name, and
account_type.

Since this endpoint returns data from classes in source code,
we needed to ensure the statement parser classes are automatically
required when rails boots. Therefore, a config/initializer was
created for this specific purpose.

In a future commit, the accounts controller will be updated to take
in an accountProvider param which will indicte the account the user
wants to add. The account instance will then be populated with values
from the ParserClass.