Update Account Controller params for accountProvider
The major update in this commit changes the way accounts are created.
Users will no longer specify a bank name or account name or account
type. Instead, they will pass in an accountProvider which is listed
in the accounts/supported. Then, the CreateAccountService will
lookup the StatementParser class corresponding to the accountProvider.
The account attributes are populated from the StatementProvider class.

We also remove the parserClass attribute from the Account attributes
that are returned to users. This parserClass attribute is not
necessary, and it's more clear to return the bank_name, name, and
account_type instead.

Once the account is created, the accountProvider cannot be changed.
The only attributes that can be updated are the statement_directory
and the active fields.

All of the tests were updated to test this new behavior.