Fix parseStatementJob tests
There was a bug in the parseStatementJob test where the ordering
of the execution would cause the "should process statements and
transactions" test to fail if it wasn't executed first. This was
happening because the ENV variable was being reset inbetween each
test which didn't actually work. So when the parser actually needed
to retreive the correct ROOT_DOC_DIR env variable, it was set to the
first value that was set during the first test. Only the "should
process statements and transactions" test makes it through the
entire parser, so if it was incorrectly set for the other tests, it
didn't matter.

To fix it, we instead use a class variable and keep the root
directory the same for all tests, and only change the account directory
between each test.