Add filepath and images to account providers
This adds the ability to return filenames for each account's stock
image. Every Statement Parser class was updated to add a new constant
for specifying the filename.

The Entity will return the relative URL path to the filename.

This also takes the opportunity to update the Schwab Checking parser
to Charles Schwab Checking to align with the industry name.