Complete add and delete account functionality
This completes the functionality to add and remove an account. The
add account is completed through a two page modal. The first page
has the user select the account, and the second page allows them to
enter details such as the user and the statement directory.

The delete account is completed by clicking on the existing account,
and then pressing delete. A warning will appear stating that all the
transactions will be deleted.