Add searchbar to crete account modal
This adds a search bar to the create account modal to search for
account providers during the first phase of adding an account.

I designed the UX such that if the user clicks an account but
decides to go back, the search query is still populated. We only
clear the search query if the modal is closed, along with resetting
all the other state.