Add filters to Accounts index
This adds filter functionality to the Accounts index API endpoint.
The current supported filters are user_ids and account_types.

The account_types are a string that matches the account_type field
in the accounts.

An account data entity is created to handle the filter logic.

Again, we convert camelCase to snake_case because API are supposed to
be in camelCase, but Ruby works in snake_case. Therefore, the controller
acts as this translation layer.

Tests were added to ensure the entity functionality.