Add Dockerfile for building frontend
Adds a Dockerfile to build the frontend next.js app. This also fixes
warnings and deprecations that appeared in the build logs. For example,
all pages containing useSearchParams must be wrapped in suspense, so
the Account and Transaction pages are now wrapped in Suspense
components. There were other warnings as well like not setting both
height and width on images, and using Image instead of img tags.

We also add an environment variable for the backend API.

This also adds the frontend service to the docker-compose file.

There is one deprecation warning that we don't have the ability to fix
because eslint uses punycode downstream. Eslint has yet to release a
version that moves off of punycode, but it's still functioning for now.