Add Nginx as SSL terminus
This adds Nginx as a service to our webapp. Nginx will be the
terminus for SSL instead of the individual Rails and Next.js
applications.

We remove the SSL configuration from Puma.rb so that Rails will
boot up using http on port 3005. Nginx will pass all traffic to
api.tincan.com to Rails.

All other traffic will go to the frontend app on port 3000.

This also adds a new nginx root directory, along with the nginx.conf,
and the directory that will store the SSL certs. The SSL certs will
never be checked in, and must be placed into the ssl directory on
the server in which this app will run.

This also updates the docker-compose file to launch the nginx service.