Temporarily use http over https
We dont yet have a domain name for this webapp. Therefor, we don't
have the ability to register an SSL cert.

I attempted to use a self-signed cert and planned to accept the
browswer warning when opening the site, but the self signed cert
breaks the XHR browser requests to the API. There is no getting
around that.

For now, since this app will only be used on the local network,
we will use http. When we register a domain, this conf file will be
updated to use SSL again.