Move account images to frontend app
This relocates the account provider images to the frontend next.js
app. They were previously being hosted on the backend app.

We also remove the supportedAccount attribute for imageFilename.
The image filename will be inferred from the accountProvider. This
makes it easier to run the app in production behind nginx. All
static images should go to the frontend app, and only API data
calls will be routed to the backend.

This removes the filename constant that was set on all of the
StatementParser classes as they are no longer necessary.