Add active scope to Account
This adds a model scope to Account for active. We'll be needing
this scope in the statement parser job to only parse statements
for active accounts.

Tests were added for this new scope, as well as a new fixture for
and inactive account. Tests in the data_entity_test were also
updated to account for the new inactive account we added to the
fixtures.