Second format for Chase Freedom statement balance
I discovered that the Chase Freedom statement has another format
for the statement balance field. This second format was added as
another regex, and if the first one doesn't detect the balance, it
falls back onto the second one.

Tests were added and existing spec tests were updated to account for
this new regex.

We also take this opportunity to update the Statement Parser job to
select "active" accounts only.