Transactions apply categorization rules
This adds logic to the Transactions model to search for a matching
categorization rule on create, and if one is found, applies the
rule's category and subcategory to the transaction.

If no rule is found on create, then it defaults to the Uncategorized
category.

We also needed to add a sync_category_to_subcategory hook so that when
the subcategory is updated, the model also applies the corresponding
category since we do not update category manually. Category is always
set from the subcategory. Users will only ever specify a subcategory
for transactions.

Tests were added to the transaction model for this logic.