Small fixes in parse statements job and tests
This adds a check to ensure there is no trailing forward slash in
the ROOT_DOC_DIR since we automatically put it in when combining
the statement directories.

This also updates the job test to remove the statement directory
instead of the account directory. This was causing the test to leave
and empty test_account directory in the project root everytime we
ran tests.