Add unprocessable entity error
Adds a custom exception class for an unprocessable entity. This
class will be used to format and render user facing errors related
to invalid parameters when creating or updating records.

The error class can take in an ActiveModel::Errors class, which is
returned by invalid models by default, or a hash. It will format
the response to be an array of objects with a field and message
attribute.