Add transform camel case params to application controller
This commit extracts the transform camel case params to the
application_controller so that every controller automatically gets
their params converted from camelCase to snake_case. This is
required because API industry standards state APIs must use
camelCase params. However, rails uses snake_case for model
attributes.

We also implement the new UnprocessableEntityError into all of the
existing controllers (except for transactions controller which is
special and we will refactor that later). Whenever there is an
unprocessable entity such as invalid model params, or an invalid
param passed in with the request, we will raise this new exception.

The create account service was refactored to incorporate the new
unprocessable entity error for an invalid account model as well as
an invalid statementParser value.

Spec tests were updated.