Add categorization rules controller and logic
This adds the categorization rules controller and all of the logic
necessary for a user to interact with the CategorizationRules
resource.

A user will create or update a CategorizationRule by passing in the
subcategory, and the category will be inferred from the
subcategory. A specific error is returned if a user passes in a
categoryId param to let them know it's not allowed.