Add BadRequestError custom exception class
This commits adds a new BadRequestError that will handle invalid
parameters for GET operations. The return code will be a 400.

The application controller was setup to rescue these exceptions
and render the correctly formatted error messages from the
exception class.

Tests were added for the application controller rescuing
BadRequestError exceptions and a backfill test for rescuing
UnprocessableEntityError exceptions.