Refactor Transactions controller and entity
This refactors the transactions controller and entity to follow the
new patten of having the params converted to snake_case by the
application controller, passing in the permitted read params to the
transaction data entity, and raising the new BadRequestError when
bad params are passed in for read operations.

We also begin to raise a UnprocessessableEntity errror if bad
params are passed in for the write operations.

The transaction data entity was refactored to handle the read param
checking and will raise BadRequestErrors if any are invalid.