Add simplecov for coverage reports
This adds the simplecov gem for test coverage reports.

However, it doesn't work when tests are parallelized. Therefore, we
disable the parallelization since the tests take less than 2
seconds to run in series. We can revisit making simplecov work with
parallelization if and when the tests take much longer.

This also installs the ostruct gem since it will no longer be
included in the standard library in a upcoming version of Ruby
(3.5.0).