Update transaction params to be camelCase
This commit updates the frontend and backend apps so that the
transaction params start to use camelCase for read and write
operations. The Typescript definitions have been updated, as well
as all references to these attributes in the frontend.

There are some auto linter format changes as well since we've
implented a front end linter that will update the typescript files
on save.