Add categorization rule and condition types and functions
This commit adds the Typescript type definitions for the
categorization rule and categorization condition. We also add the
functions to make API calls to the backend for fetching, creating,
updating and deleting these two resources.

These are not yet used in the frontend app, but a future commit
will build a UI around these resources.