Update db:seeds
This commit updates db/seeds so that we are seeding the database
with all of the categories, subcategories, current accounts and a
few categorization rules and conditions.

This removes the transaction seeds since the statement parser job
is working, and we can easily parse the statements to load
transactions. We no longer need to create fake transactions.