Add auto-categorization rule management page
This adds the initial auto-categorization rule management page. As
of now the only functionality is to view the rules if any exist.

If there are no rules, a box will display that states there aren't
any rules yet and that they should add one, but the button isn't
implemented yet.

There is also an "about" dropdown that explains what
auto-categorization rules are.