Generalize Search component
This generalizes the Search component so that it can be used
across more pages instead of only the transactions page.

We take in props such as the onSearch function which is called when
text is changed in the search bar, as well as a value string.

We also create a new directory src/components which will house
the shared components of the app.