Remove font specifications from filter components
This removes the font setting code inside the transactions filter
and accounts filter components. The parent page is setting the font
for all components on the page, and so setting the font is no
longer necessary inside the filter components.

We also change the font size of the filter header from 2xl to xl
for the transaction filter to match the account filter.

There are also some linter formatting changes that were
auto-corrected.