Add filter bar to categorization rule page
This adds a filter bar section to the categorization rule page. The
functionality is not yet implemented, but the UI is displaying
correctly.

We also refactor the layout of the page so that the entire content
panel has a max width of 4xl. All other components within the
content panel are sized to full width, and the parent div will
restrict it to 4xl and center the content.

We also add logic to hide the text inside the action buttons when
the window becomes smaller than md size.

There is also a small update to remove an unused import in the
categorization dropdown component.