Generalize Search component
This refactors the Search component so that it can be used by other
components. We add the functionality to clear the search text by
giving access to the value rather than the default value.

I also noticed that the transactions page flickers when search text
is entered into the search box. This is because if the transactions
are loading, we set the page UI to just say "loading..." which
makes it look like it's re-rendering. Instead, we should show a
suspense table object so that the entire page doesn't appear to
reload. For now, we simply remove the check for
isLoadingTransactions.