Refactor Accounts modals
This refactors the Accounts page modals to split out the create
screens into different components. We create a subdirectory for
each modal, the create account modal and the update account modal.
The update modal only has one screen.

We also update the account grouping logic in the accounts table so
that the accounts are sorted by name inside the bank and user
groups.

We also shift the call to the group function to happen in a
useEffect call. This way, the group logic is only executed when the
accounts array changes. Prior to this it was executing on ever
render, even when accounts didn't change.

We also add some improvements to the UI buttons. For example, the
update modal now has icons for the delete and close to quickly
identify the correct button. Additionally, the close will have a
confirmation if the statement directory is unsaved.