Rubocop refactors for transactions controller
This fixes all rubocop offenses for the transactions controller and
its tests. One offense was that the test controller was too long,
so we split it up into separate test classes based on the
controller endpoint it was testing: index, update, create, destroy.

We also upgrade rubocop, add rubocop-rails, and create a
rubocop.yml config file to ignore certain offenses.