Fix bug in categorization_rule match? logic
This fixes a bug in the categorization_rule match? logic when a
rule has no conditions associated with it. An empty rule with no
conditions should always return false when checking for a match.
However, the all? returned true when there were no items in the
iterator.

Tests were added for the various match? cases.

This also fixes some rubocop offenses.