Fix bug in transactions controller call to entity
This fixes a bug in which the transactions data entity method
get_data was refactored to .data, however the controller was never
updated. Tests weren't catching this because we were mocking the
entity call in the transactions controller tests. So, we update the
controller to call the correct method. We also update the tests so
that we stop mocking the entity, and ensure that the correct
parameters are passed to the entity and call the original.