Refactor categorization rules to include conditions
This is a major refactor of the categorization rule resource.

First and foremost, we update the interface to include conditions
within the categorization rules. On index, we will display the list
of conditions within the rule. On create and update, we allow
providing conditions to the rule.

If conditions are not provided on update, they will remain
unchanged. If conditions are provided on update, they will be
overwritten. There is no merging of conditions.

This also updates the validation logic of the condition model. We
are more specific in the error messages for when fields are invalid
or missing.

This commit also includes a bunch of refactors to fix rubocop
offenses.